+++ backend/src/main/java/com/nutritrack/dto/FoodUpdateRequest.java
@@ -20,10 +20,10 @@ public class FoodUpdateRequest {
 
     @Data
     public static class NutritionRequest {
-        private float calories;
-        private float protein;
-        private float carbs;
-        private float fat;
+        private int calories;
+        private int protein;
+        private int carbs;
+        private int fat;
     }
 
     @Data
